[IMP] mail:tests: unskip 'chat window should scroll to the newly posted message just after posting it'
